10 - funzionalità di delete user + interests

+++ usermanager/app.py
@@ -7,6 +7,9 @@ import service_pb2
 import service_pb2_grpc
 import threading
 from sqlalchemy import select
+from datetime import datetime, timedelta
+import schedule
+import time
 app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = os.environ.get('DATABASE_URL')
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
@@ -16,17 +19,27 @@ class User(db.Model):
     email = db.Column(db.String(255), primary_key=True)
     name = db.Column(db.String(100), nullable=False)
     surname = db.Column(db.String(100), nullable=False)
+    is_active = db.Column(db.Boolean, default=True, nullable=False)
+    deleted_at = db.Column(db.DateTime, nullable=True)
 class UserManagerService(service_pb2_grpc.UserManagerServiceServicer):
-    def ValidateEmail(self, request, context):
+    def CheckUserStatus(self, request, context):
         email=request.email
         print(f"Server: Ricevuta richiesta con utente: {email}", file=os.sys.stderr)
         with app.app_context():
             stmt = select(User).where(User.email == email)
             user = db.session.execute(stmt).scalar_one_or_none()
-            if user:
-                return service_pb2.Response(validate=True)
+            if user is None:
+                return service_pb2.Response(
+                    status=service_pb2.UserStatus.NOT_FOUND
+                )
+            if user.is_active:
+                return service_pb2.Response(
+                    status=service_pb2.UserStatus.ACTIVE
+                )
             else:
-                return service_pb2.Response(validate=False)
+                return service_pb2.Response(
+                    status=service_pb2.UserStatus.DELETED
+                )
 def serve():
     try:
         #print(f"Server started at :{os.environ.get('GRPC_PORT')}")
@@ -45,7 +58,7 @@ def home():
 @app.route('/users', methods=['GET'])
 def users():
     users= db.session.execute(db.select(User).order_by(User.email)).scalars()
-    users_list = [{"email": user.email, "name": user.name, "surname": user.surname} for user in users]
+    users_list = [{"email": user.email, "name": user.name, "surname": user.surname, "is_active": user.is_active, "deleted_at": user.deleted_at} for user in users]
     return jsonify({"users": users_list}), 200
 @app.route('/add_user', methods=['POST'])
 def add_user():
@@ -56,11 +69,38 @@ def add_user():
 @app.route('/delete_user', methods=['POST'])
 def delete_user():
     user = User.query.filter_by(email=request.form['email']).first()
-    db.session.delete(user)
+    user.is_active = False
+    user.deleted_at = datetime.utcnow()
     db.session.commit()
-    return jsonify({"message": "User deleted successfully"}), 200
+    return jsonify({"message": "User deleted successfully (Soft Delete)"}), 200
+def hard_delete_user():
+    with app.app_context():
+        deleted_count = 0
+        threshold_time = datetime.utcnow() - timedelta(hours=24)
+        #print(f"Hard delete {threshold_time}", file=os.sys.stderr)
+        users= db.session.execute(db.select(User).order_by(User.email)).scalars()
+        for user in users:
+            if user.is_active == False and user.deleted_at > threshold_time:
+                db.session.delete(user)
+                deleted_count += 1
+        # deleted_count = db.session.query(User).filter(
+        #     User.is_active == False,
+        #     User.deleted_at <= threshold_time
+        # ).delete(synchronize_session=False)
+        db.session.commit()
+        if deleted_count > 0:
+            print(f"[HARD DELETE] Rimossi definitivamente {deleted_count} utenti obsoleti.", file=os.sys.stderr)
+def run_deletion():
+    schedule.every(20).seconds.do(hard_delete_user)
+    #print(f"[HARD DELETE]", file=os.sys.stderr)
+    hard_delete_user()
+    while True:
+        schedule.run_pending()
+        time.sleep(1)
 if __name__ == '__main__':
     grpc_thread = threading.Thread(target=serve, daemon=True)
     grpc_thread.start()
+    scheduler_thread = threading.Thread(target=run_deletion, daemon=True)
+    scheduler_thread.start()
     #print("mainDEBUG: GRPC_PORT =", os.environ.get("GRPC_PORT"))
     app.run(host='0.0.0.0', port=5000, debug=False)
